Added Batch insert/delete.

+++ app/src/androidTest/java/io/github/dnloop/noteapp/data/NoteDaoTest.kt
@@ -198,7 +198,7 @@ class NoteDaoTest {
         val note = Note()
         // insert multiple fields
         for (i in 1..5){
-            note.id = i
+            note.id = i.toLong()
             noteDao.insert(note)
         }
         // check assertion
